refator: weapon & skill 의 상호 데미지 연결 구조 리팩토링
다음 작업 목표 VFX 및 Projectile 클래스 리팩토링

+++ Assets/Script/Enemies/EnemyHealth.cs
@@ -4,10 +4,10 @@ using UnityEngine;
 
 public class EnemyHealth : MonoBehaviour
 {
-    [SerializeField] private int maxHealth = 3;
+    [SerializeField] private float maxHealth = 3;
     [SerializeField] private GameObject deathVFXPrefab;
     [SerializeField] private float knockBackThrust = 15f;
-    private int curHealth;
+    private float curHealth;
     private Knockback knockback;
     private Flash flash;
     private void Awake()
@@ -19,7 +19,7 @@ public class EnemyHealth : MonoBehaviour
     {
         curHealth = maxHealth;
     }
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         curHealth -= damage;
         knockback.GetKnockedBack(PlayerController.Instance.transform, knockBackThrust);
